third_party/WebKit/LayoutTests/inspector-protocol/input/eventTimestamp-expected.txt

Tests timestamps in multiple input domain methods.
-----Event-----
type: keydown
-----Event-----
type: keydown
-----Event-----
type: mousedown
-----Event-----
type: mousedown
-----Event-----
type: touchstart
-----Event-----
type: touchstart
-----Verify-----
Received 6 timestamps
timeStamps offsets is as expected.
timeStamps offsets is as expected.
timeStamps offsets is as expected.
timeStamps offsets is as expected.
timeStamps offsets is as expected.
timeStamps offsets is as expected.
